requirements.txt
flask
google-generativeai
Pillow
python-dotenv
geopy